pyproject.toml
[project]
name = "flywheel-gear-dev-mcp"
version = "0.1.0"
description = "MCP server providing Flywheel gear development documentation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "markdownify>=0.12.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
flywheel-gear-mcp = "flywheel_gear_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flywheel_gear_mcp"]
